examples/DockerCompose/flask-app/requirements.txt
Flask
psycopg2-binary  # For PostgreSQL integration
